+++ example/lib/models/user_model.dart
@@ -3,6 +3,16 @@ import 'package:mvc_rocket/mvc_rocket.dart';
 import 'user_submodel/address_submodel.dart';
 import 'user_submodel/company_submodel.dart';
 
+const String userIdField = "id";
+const String userNameField = "name";
+const String userUsernameField = "username";
+const String userEmailField = "email";
+const String userAddressField = "address";
+const String userPhoneField = "phone";
+const String userWebsiteField = "website";
+const String userCompanyField = "company";
+const String userImageField = "image";
+
 class User extends RocketModel<User> {
   int? id;
   String? name;
@@ -13,16 +23,6 @@ class User extends RocketModel<User> {
   String? website;
   Company? company;
   String? image;
-
-  String idVar = "id";
-  String nameVar = "name";
-  String usernameVar = "username";
-  String emailVar = "email";
-  String addressVar = "address";
-  String phoneVar = "phone";
-  String websiteVar = "website";
-  String companyVar = "company";
-  String imageVar = "image";
 
   User(
       {this.id,
@@ -40,30 +40,55 @@ class User extends RocketModel<User> {
 
   @override
   void fromJson(covariant Map<String, dynamic> json, {bool isSub = false}) {
-    id = json['id'] ?? id;
-    name = json['name'] ?? name;
-    username = json['username'] ?? username;
-    email = json['email'] ?? email;
-    address!.fromJson(json['address'] ?? address!.toJson(), isSub: isSub);
-    phone = json['phone'] ?? phone;
-    image = json['image'] ?? image;
-    website = json['website'] ?? website;
-    company!.fromJson(json['company'] ?? company!.toJson(), isSub: isSub);
+    id = json[userIdField] ?? id;
+    name = json[userNameField] ?? name;
+    username = json[userUsernameField] ?? username;
+    email = json[userEmailField] ?? email;
+    address!
+        .fromJson(json[userAddressField] ?? address!.toJson(), isSub: isSub);
+    phone = json[userPhoneField] ?? phone;
+    image = json[userImageField] ?? image;
+    website = json[userWebsiteField] ?? website;
+    company!
+        .fromJson(json[userCompanyField] ?? company!.toJson(), isSub: isSub);
     super.fromJson(json, isSub: isSub);
+  }
+
+  void updateFields({
+    int? idField,
+    String? nameField,
+    String? usernameField,
+    String? emailField,
+    Address? addressField,
+    String? phoneField,
+    String? websiteField,
+    Company? companyField,
+    String? imageField,
+  }) {
+    id = idField ?? id;
+    name = nameField ?? name;
+    username = usernameField ?? username;
+    email = emailField ?? email;
+    address = addressField ?? address;
+    phone = phoneField ?? phone;
+    image = imageField ?? image;
+    website = websiteField ?? website;
+    company = companyField ?? company;
+    rebuildWidget();
   }
 
   @override
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> data = {};
-    data['id'] = id;
-    data['name'] = name;
-    data['username'] = username;
-    data['email'] = email;
-    data['address'] = address!.toJson();
-    data['phone'] = phone;
-    data['website'] = website;
-    data['company'] = company!.toJson();
-    data['image'] = image;
+    data[userIdField] = id;
+    data[userNameField] = name;
+    data[userUsernameField] = username;
+    data[userEmailField] = email;
+    data[userAddressField] = address!.toJson();
+    data[userPhoneField] = phone;
+    data[userWebsiteField] = website;
+    data[userCompanyField] = company!.toJson();
+    data[userImageField] = image;
 
     return data;
   }