+++ lib/src/mc_model.dart
@@ -28,9 +28,12 @@ abstract class RocketModel<T> extends RocketListenable {
   }
 
   /// التقاط الخطأ
-  void setException(RocketException response) {
-    exception = response;
+  void setException(RocketException exception) {
+    exception = exception;
     state = RocketState.failed;
+    if (enableDebug) {
+      log("[MVCR]> Failed details :${exception.toString()}");
+    }
   }
 
   set loadingChecking(bool value) {
@@ -64,8 +67,17 @@ abstract class RocketModel<T> extends RocketListenable {
     }
   }
 
+  /// update fields based on fromJson method
+  void updateFieldsByMap(Map<String, dynamic>? json) {
+    try {
+      fromJson(json);
+    } catch (e) {
+      setException(RocketException(exception: e.toString()));
+    }
+  }
+
   /// json من النماذج الى بيانات
-  Map<String, dynamic> toJson();
+  Map<String, dynamic> toJson() => {};
 
   DateTime _time = DateTime.now();
 
@@ -74,12 +86,12 @@ abstract class RocketModel<T> extends RocketListenable {
     if (enableDebug) {
       if (state == RocketState.loading) {
         _time = DateTime.now();
-        log('${T.toString()} : ${state.name}');
+        log('[MVCR]> ${T.toString()} : ${state.name}');
       }
 
       if (state == RocketState.done || state == RocketState.failed) {
         int dur = DateTime.now().difference(_time).inMilliseconds;
-        log('${T.toString()} : ${state.name} in $dur ms');
+        log('[MVCR]> ${T.toString()} : ${state.name} in $dur ms');
       }
     }